@@ -3,22 +3,33 @@ plugins {
     alias(libs.plugins.kotlinAndroid)
 }
 
-val androidCompileSDK : String by project
-val androidMinSDK : String by project
+kotlin {
+    jvmToolchain {
+        this.languageVersion.set(JavaLanguageVersion.of(libs.versions.jvmTarget.get()))
+    }
+}
 
 android {
-    compileSdk = androidCompileSDK.toInt()
+
+    compileSdk = libs.versions.android.compileSdk.get().toInt()
+    namespace = "org.koin.android.compat"
     defaultConfig {
-        minSdk = androidMinSDK.toInt()
+        minSdk = libs.versions.android.minSdk.get().toInt()
     }
     buildFeatures {
         buildConfig = false
     }
     compileOptions {
-        sourceCompatibility = JavaVersion.VERSION_1_8
+        sourceCompatibility = JavaVersion.toVersion(libs.versions.jvmTarget.get())
+    }
+    publishing {
+        singleVariant("release") {
+            withSourcesJar()
+            withJavadocJar()
+        }
     }
     kotlinOptions {
-        jvmTarget = JavaVersion.VERSION_1_8.toString()
+        jvmTarget = libs.versions.jvmTarget.get()
     }
 }
 
@@ -32,4 +43,4 @@ val sourcesJar: TaskProvider<Jar> by tasks.registering(Jar::class) {
     from(android.sourceSets.map { it.java.srcDirs })
 }
 
-apply(from = file("../../gradle/publish-android.gradle.kts"))
+//apply(from = file("../../gradle/publish-android.gradle.kts"))
